Consolidate JavaScript HTTP module specifications (#10628)
* Rename RCTNetworkingWinShared.js as RCTNetworking.windows.js

* Make RCTNetworking.win32.js compatible with RCTNetworking.windows.js

* Change files

* Restore setupDeveloperTools.windesktop.js

* Remove duplicate overrides
