+++ change/@office-iss-react-native-win32-5d4d2e75-73de-4731-96a7-622b462c7d8a.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "Consolidate JavaScript HTTP module specifications",
+  "packageName": "@office-iss/react-native-win32",
+  "email": "julio.rocha@microsoft.com",
+  "dependentChangeType": "patch"
+}
