+++ change/react-native-platform-override-563b2a70-909c-4494-a291-9bf419ee59fd.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Consolidate JavaScript HTTP module specifications",
+  "packageName": "react-native-platform-override",
+  "email": "julio.rocha@microsoft.com",
+  "dependentChangeType": "patch"
+}
