+++ change/react-native-windows-1fcfe09c-4814-4557-8df4-7e9d39f6a908.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "Consolidate JavaScript HTTP module specifications",
+  "packageName": "react-native-windows",
+  "email": "julio.rocha@microsoft.com",
+  "dependentChangeType": "patch"
+}
